+++ src/run.py
@@ -1,4 +1,27 @@
-#!/usr/bin/env python3
+#!/bin/sh
+""":"
+if type python3 >/dev/null 2>/dev/null; then
+    subversion=$(python3 --version | cut -d. -f2)
+    if test "$subversion" -ge 6; then
+        exec python3 "$0" "$@"
+    else
+        echo "Default Python version is lower than Python 3.6. Trying to force the version." >&2
+        for avail_subversion in $(whereis python3 | grep -o [0-9]\.[0-9] | sort | uniq | cut -d. -f2); do
+            if test "$avail_subversion" -ge 6; then
+                if type python3."$avail_subversion" >/dev/null 2>/dev/null; then
+                    echo "Forcing execution using Python 3.$avail_subversion."
+                    exec python3."$avail_subversion" "$0" "$@"
+                fi
+            fi
+        done
+        echo "No suitable version found. Exiting." >&2
+        exit 2;
+    fi
+else
+    echo "Unable to find Python 3. Exiting." >&2
+    exit 1;
+fi
+"""
 # run.py
 """ROS1 Entrypoint for 'ng_trajectory_ros'.
 """
@@ -21,7 +44,10 @@ def main():
 
 
     # Load configuration
-    node.load_config()
+    try:
+        node.load_config()
+    except Exception as e:
+        print (e)
 
 
     # Function spin() simply keeps python from exiting until this node is stopped.